feat: create working temp lora code
